move calculartions to client
